backend/requirements.txt
faiss-cpu
open_clip_torch
torch
Pillow

